refactor(insights): move query modules to api/queries/ for consistency
Standardize query file locations by moving insights queries from
insights/queries/ into api/queries/ alongside existing query modules.
Files renamed to insights_neo4j_queries.py and insights_pg_queries.py
to avoid collision with existing neo4j_queries.py.
