--- insights/queries/__init__.py
@@ -1,1 +1,0 @@
-"""Query modules for insights computations."""
